refactor: 주문 취소 엔티티 common 으로 이동
