+++ commerce/src/main/java/store/_0982/commerce/application/order/OrderPaymentProcessorService.java
@@ -11,6 +11,7 @@ import store._0982.commerce.domain.grouppurchase.GroupPurchaseRepository;
 import store._0982.commerce.domain.order.*;
 import store._0982.commerce.exception.CustomErrorCode;
 import store._0982.common.domain.grouppurchase.GroupPurchase;
+import store._0982.common.domain.order.CanceledOrder;
 import store._0982.common.domain.order.Order;
 import store._0982.common.domain.order.OrderStatus;
 import store._0982.common.domain.order.PaymentMethod;
