+++ commerce/src/main/java/store/_0982/commerce/infrastructure/order/CanceledOrderJpaRepository.java
@@ -3,8 +3,8 @@ package store._0982.commerce.infrastructure.order;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.commerce.domain.order.CancelStatus;
-import store._0982.commerce.domain.order.CanceledOrder;
+import store._0982.common.domain.order.CancelStatus;
+import store._0982.common.domain.order.CanceledOrder;
 
 import java.time.OffsetDateTime;
 import java.util.Collection;
