+++ common/src/main/java/store/_0982/common/domain/order/CanceledOrder.java
@@ -1,4 +1,4 @@
-package store._0982.commerce.domain.order;
+package store._0982.common.domain.order;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -6,9 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-import store._0982.commerce.exception.CustomErrorCode;
-import store._0982.common.domain.order.PaymentMethod;
-import store._0982.common.exception.CustomException;
 import store._0982.common.kafka.dto.OrderCanceledEvent;
 
 import java.time.OffsetDateTime;
@@ -29,9 +26,6 @@ public class CanceledOrder {
 
     @Column(name = "member_id", nullable = false, updatable = false)
     private UUID memberId;
-
-    @Column(name = "seller_id", nullable = false, updatable = false)
-    private UUID sellerId;
 
     @Column(name = "original_paid_amount", nullable = false)
     private Long originalPaidAmount;    // 취소 전 실제 결제액
@@ -85,34 +79,18 @@ public class CanceledOrder {
 
     public void markCompleted() {
         this.returnedAt = OffsetDateTime.now();
-        this.status = CancelStatus.COMPLETED;
-    }
-
-    public void markApproved() {
-        if (this.status != CancelStatus.PENDING) {
-            throw new CustomException(CustomErrorCode.INVALID_CANCEL_STATUS);
-        }
-        this.status = CancelStatus.APPROVED;
-    }
-
-    public void markRejected() {
-        if (this.status != CancelStatus.PENDING) {
-            throw new CustomException(CustomErrorCode.INVALID_CANCEL_STATUS);
-        }
-        this.status = CancelStatus.REJECTED;
+        status = CancelStatus.COMPLETED;
     }
 
     public static CanceledOrder createCanceledOrder(
             UUID orderId,
             UUID memberId,
-            UUID sellerId,
             long originalPaidAmount,
             long cancelFeeAmount,
             long shippingFeeAmount,
             long refundAmount,
             String policyId,
             String policySnapshot,
-            CancelStatus status,
             CancelReason reason,
             String detailReason,
             String idempotencyKey,
@@ -120,14 +98,12 @@ public class CanceledOrder {
         return new CanceledOrder(
                 orderId,
                 memberId,
-                sellerId,
                 originalPaidAmount,
                 cancelFeeAmount,
                 shippingFeeAmount,
                 refundAmount,
                 policyId,
                 policySnapshot,
-                status,
                 reason,
                 detailReason,
                 idempotencyKey,
@@ -137,14 +113,12 @@ public class CanceledOrder {
     private CanceledOrder(
             UUID orderId,
             UUID memberId,
-            UUID sellerId,
             long originalPaidAmount,
             long cancelFeeAmount,
             long shippingFeeAmount,
             long refundAmount,
             String policyId,
             String policySnapshot,
-            CancelStatus status,
             CancelReason reason,
             String detailReason,
             String idempotencyKey,
@@ -152,14 +126,13 @@ public class CanceledOrder {
         this.canceledOrderId = UUID.randomUUID();
         this.orderId = orderId;
         this.memberId = memberId;
-        this.sellerId = sellerId;
         this.originalPaidAmount = originalPaidAmount;
         this.cancelFeeAmount = cancelFeeAmount;
         this.shippingFeeAmount = shippingFeeAmount;
         this.refundAmount = refundAmount;
         this.policyId = policyId;
         this.policySnapshot = policySnapshot;
-        this.status = status;
+        this.status = CancelStatus.REQUESTED;
         this.reason = reason;
         this.detailReason = detailReason;
         this.idempotencyKey = idempotencyKey;
